requirements.txt
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
click==8.1.3
configparser==5.0.2
cycler==0.10.0
docker-pycreds==0.4.0
filelock==3.13.1
fsspec==2023.10.0
gitdb==4.0.7
GitPython==3.1.16
graphviz==0.20.1
huggingface-hub==0.19.4
idna==2.10
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.3.1
MarkupSafe==2.1.1
matplotlib==3.4.2
memory-profiler==0.61.0
ninja==1.11.1.1
numpy==1.20.3
packaging==23.2
pandas==1.2.4
pathtools==0.1.2
Pillow==8.2.0
promise==2.3
protobuf==3.17.0
psutil==5.8.0
pyg-lib==0.1.0+pt113cpu
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
safetensors==0.4.0
scikit-learn==1.2.0
scipy==1.10.0
sentry-sdk==1.1.0
setproctitle==1.3.2
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
subprocess32==3.5.4
tensorboardX==2.2
threadpoolctl==3.1.0
timm==0.9.10
torch==1.10.1
torch-cluster==1.6.0+pt113cpu
torch-geometric==2.2.0
torch-scatter==2.1.0+pt113cpu
torch-sparse==0.6.16+pt113cpu
torch-spline-conv==1.2.1+pt113cpu
torchaudio==0.10.1
torchvision==0.9.1
torchviz==0.0.2
tqdm==4.62.3
typing-extensions==3.10.0.0
urllib3==1.26.4
wandb==0.14.0
